tests/wpt/tests/svg/animations/smil-values-trailing-semicolon.html

<!DOCTYPE html>

<title>Trailing semicolons in SMIL values attribute should produce valid path</title>
<meta charset="utf-8">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link rel="help" href="https://svgwg.org/specs/animations/#ValuesAttribute">

<svg viewbox="0 0 100 100">
  <path fill="#1EB287">
    <animate attributeName="d"
             dur="1s"
             repeatCount="indefinite"
             values="M 0,0 L 50, 0 L 100,100 Z;
                     M 100,0 L 0, 100 L 50, 0 Z; "/>
  </path>
</svg>

<script>
  async_test(function(t) {
    window.onload = t.step_func(function() {
      window.requestAnimationFrame(t.step_func_done(function() {
        let path = document.querySelector("path");
        let dValue = window.getComputedStyle(path).d;
        assert_not_equals(dValue, "none", "d property should not be 'none'");
        assert_regexp_match(dValue, /^path\(/, "d property should be a valid path value");
      }));
    });
  });
</script>
